Changed all Vec3i to Vec3f

+++ ascii.go
@@ -11,57 +11,57 @@ type Vec3 interface {
 	Add(rhs Vec3)
 } 
 
-type Vec3i struct {
-	X int
-	Y int
-	Z int
+type Vec3f struct {
+	X float64
+	Y float64
+	Z float64
 }
 
-func (v *Vec3i) Add(rhs *Vec3i) *Vec3i {
+func (v *Vec3f) Add(rhs *Vec3f) *Vec3f {
 	v.X += rhs.X
 	v.Y += rhs.Y
 	v.Z += rhs.Z
 	return v
 }
 
-func (v *Vec3i) Dot(rhs *Vec3i) int {
+func (v *Vec3f) Dot(rhs *Vec3f) float64 {
 	return v.X*rhs.X + v.Y*rhs.Y + v.Z*rhs.Z
 }
 
-func (v *Vec3i) Norm() float64 {
-	return math.Sqrt(float64(v.Dot(v)))
+func (v *Vec3f) Norm() float64 {
+	return math.Sqrt(v.Dot(v))
 }
 
-func (v *Vec3i) ScalarMul(a int) *Vec3i {
+func (v *Vec3f) ScalarMul(a float64) *Vec3f {
 	v.X *= a
 	v.Y *= a
 	v.Z *= a
 	return v
 }
 
-func (v *Vec3i) Sub(rhs *Vec3i) *Vec3i {
+func (v *Vec3f) Sub(rhs *Vec3f) *Vec3f {
 	return v.Add(rhs.ScalarMul(-1))
 }
 
-func generateCube(corner *Vec3i, sideLength int) []Vec3i {
-	pts := make([]Vec3i, 0, (sideLength+1)*(sideLength+1)*(sideLength+1))
+func generateCube(corner *Vec3f, sideLength int) []Vec3f {
+	pts := make([]Vec3f, 0, (sideLength+1)*(sideLength+1)*(sideLength+1))
 	for x := 0; x <= sideLength; x++ {
 		for y := 0; y <= sideLength; y++ {
 			for z := 0; z <= sideLength; z++ {
-				pts = append(pts, Vec3i{corner.X+x, 
-										corner.Y+y,
-										corner.Z+z})
+				pts = append(pts, Vec3f{corner.X+float64(x), 
+										corner.Y+float64(y),
+										corner.Z+float64(z)})
 			}
 		}
 	}
 	return pts
 }
 
-func render(pts []Vec3i, lightSource Vec3i, lightIntensity, xs, ys, zs int) [][]float64 {
+func render(pts []Vec3f, lightSource Vec3f, lightIntensity, xs, ys, zs int) [][]float64 {
 	xs++
 	ys++
 	zs++
-	camera := Vec3i{xs/2, ys/2, 0}
+	camera := Vec3f{float64(xs/2), float64(ys/2), 0.0}
 	//Initialize two 2D arrays for the screen and for the z-buffer.
 	pixUnderlying  := make([]float64, xs*ys)
 	zbufUnderlying := make([]float64, xs*ys)
@@ -74,8 +74,8 @@ func render(pts []Vec3i, lightSource Vec3i, lightIntensity, xs, ys, zs int) [][]
 
 	for _, pt := range pts {
 		x, y, z := pt.X-camera.X, pt.Y-camera.Y, pt.Z-camera.Z
-		xp := zs/z * x + camera.X
-		yp := zs/z * y + camera.Y
+		xp := int(float64(zs)/z * x + camera.X)
+		yp := int(float64(zs)/z * y + camera.Y)
 		inv_z := 1.0/float64(z)
 		if xp>0 && xp<xs && 
 		   yp>0 && yp<ys && 
@@ -88,10 +88,10 @@ func render(pts []Vec3i, lightSource Vec3i, lightIntensity, xs, ys, zs int) [][]
 			//should be surface normal...
 			d := pt
 			d.Sub(&lightSource)
-			t := Vec3i{x, y, z}
+			t := Vec3f{x, y, z}
 			fmt.Println(d)
 			light := d.Dot(&t)
-			pix[yp][xp] = float64(light * lightIntensity) / (d.Norm()*t.Norm())
+			pix[yp][xp] = light * float64(lightIntensity) / (d.Norm()*t.Norm())
 		}
 	}
 	return pix
@@ -116,14 +116,14 @@ func display(pix [][]float64) {
 }
 
 func main() {
-	pts := generateCube(&Vec3i{10, 10, 8}, 3)
-	pix := render(pts, Vec3i{0, -2, 0}, 12, 15, 15, 15)
+	pts := generateCube(&Vec3f{10, 10, 8}, 3)
+	pix := render(pts, Vec3f{0, -2, 0}, 12, 15, 15, 15)
 	display(pix)
 
 	time.Sleep(3*time.Second)
 
-	pts = generateCube(&Vec3i{3, 10, 8}, 3)
-	pix = render(pts, Vec3i{0, -2, 0}, 12, 15, 15, 15)
+	pts = generateCube(&Vec3f{3, 10, 8}, 3)
+	pix = render(pts, Vec3f{0, -2, 0}, 12, 15, 15, 15)
 	display(pix)
 }
